+++ lib/device/StripDescription.cpp
@@ -3,7 +3,7 @@
 #include <functional>
 #include <numeric>
 
-StripDescription::StripDescription(uint8_t led_count,
+StripDescription::StripDescription(uint16_t led_count,
                                    std::vector<StripFlag> flag_list)
     : led_count(led_count),
       flags(std::accumulate(flag_list.begin(), flag_list.end(), 0,
