+++ lib/effect/Effect.hpp
@@ -13,7 +13,7 @@ class Effect {
   virtual ~Effect(){};
 
   /** Gets the value of a specific LED at a specific time. */
-  virtual CRGB GetRGB(uint8_t led_index, uint32_t time_ms,
+  virtual CRGB GetRGB(uint16_t led_index, uint32_t time_ms,
                       const StripDescription &strip,
                       RadioPacket *setEffectPacket) const = 0;
