+++ lib/effect/RainbowEffect.cpp
@@ -2,7 +2,7 @@
 
 RainbowEffect::RainbowEffect() : Effect() {}
 
-CRGB RainbowEffect::GetRGB(uint8_t led_index, uint32_t time_ms,
+CRGB RainbowEffect::GetRGB(uint16_t led_index, uint32_t time_ms,
                            const StripDescription &strip,
                            RadioPacket *setEffectPacket) const {
   uint8_t v = strip.FlagEnabled(Bright) ? 255 : 128;
